% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_indices.R
\name{agg_ind}
\alias{agg_ind}
\title{Function to obtain aggregated (mean and variance) distinctiveness or
scarcity per species and community}
\usage{
agg_ind(species, com, ind)
}
\arguments{
\item{species}{a character vector counting all the species in each community}

\item{com}{a character vector giving a list of the communities of all
combinations (see ind)}

\item{ind}{a distinctiveness or scarcity vector of each species/community
association, ind is obtained thanks to distinctiveness or scarcity function}
}
\value{
Four vector with attributed names:
mean_ind_com contains the index' mean per community
mean_ind_sp contains the index' mean per species
var_ind_com contains the index' variance per community
var_ind_sp contains the index' variance per species
skew_ind_com contains the index' skewness per community
skew_ind_sp contains the index' skewness per species
}
\description{
Arguments:
}
\details{
com, a character vector combining all communities, each community is repeated
by the number of species it contains
}

