% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity_dimensions.R
\name{combination_trait_dist}
\alias{combination_trait_dist}
\title{Compute Multiple distance matrices from a single trait table}
\usage{
combination_trait_dist(traits_table, ...)
}
\arguments{
\item{traits_table}{a data.frame of traits with species in row and traits in
columns, \strong{row names} should be \strong{species names},}

\item{...}{additional arguments supplied to \code{\link[=compute_dist_matrix]{compute_dist_matrix()}}}
}
\value{
A list of functional distance matrices, one for each provided trait
plus an additional matrix for all traits taken altogether
}
\description{
Internal function to compute combinations of distance matrices from a
data.frame of traits, using \code{\link[=compute_dist_matrix]{compute_dist_matrix()}}.
}
