% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_matrix.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\alias{distance_matrix}
\title{Functional Distance Matrix}
\usage{
compute_dist_matrix(traits_table, metric = "gower")
}
\arguments{
\item{traits_table}{A data frame of traits with species in row and traits in
columns, row names should be species names,}

\item{metric}{character vector in list \code{"c('gower', 'manhattan',
euclidean')"} defining the type of distance to use
(see\code{\link[cluster]{daisy}}).}
}
\value{
A Gower's distance matrix (Gower's 1961), columns and rows get same
    names as \code{trait_table} row names. Be sure to give them consistent
    row names.
}
\description{
Wrapper for \code{"daisy()"} function in \code{"cluster"} package, to compute
distance matrix of trait between each pair of species present in given
\code{"traits_table"}, each row represents a species and each column a trait.
To be able to compute other distances trait table must have species as row
names.
}
\examples{
set.seed(1)  # For reproducibility
trait = data.frame(sp = paste("sp", 1:5), trait_1 = runif(5),
   trait_2 = as.factor(c("A", "A", "A", "B", "B")))

rownames(trait) = trait$sp

dist_mat = compute_dist_matrix(trait[, -1])

}

