% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctiveness.R
\name{distinctiveness}
\alias{distinctiveness}
\title{Distinctiveness}
\usage{
distinctiveness(com_table, species, com, abund = NULL, dist_matrix)
}
\arguments{
\item{com_table}{a data frame of the species in the regional pool.}

\item{com}{a character vector, indicating the column name of communities
names.}

\item{abund}{optional character vector indicating the column name of the
relative abundances of species, if provided, weight distinctiveness by
relative abundances.}
}
\value{
a table similar to \code{com_table} with an added column \eqn{D_i}
}
\description{
Compute distinctiveness for several communities.
}

