% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctiveness.R
\name{distinctiveness_stack}
\alias{distinctiveness_stack}
\title{Table Functional Distinctiveness}
\usage{
distinctiveness_stack(com_df, sp_col, com, abund = NULL, dist_matrix)
}
\arguments{
\item{com_df}{a stacked (or tidy) data.frame from a single community with
each row representing a species}

\item{sp_col}{a character vector the name of the species column in
\code{com_df}}

\item{com}{a character vector, indicating the column name of communities
names.}

\item{abund}{a character vector indicating the name of the column containing
relative abundances values}

\item{dist_matrix}{a functional distance matrix as given by
\code{compute_dist_matrix()}}
}
\value{
the same data.frame with the additional \strong{Di} column giving
   functional distinctiveness values for each species
}
\description{
Compute Functional Distinctiveness for several communities, from a data.frame
of communities, with one column for species identity, one for community
identity and an optional one for relative abundances.
}
\examples{
data("aravo", package = "ade4")

# Example of trait table
tra = aravo$traits[, c("Height", "SLA", "N_mass")]
# Distance matrix
dist_mat = compute_dist_matrix(tra)

# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe); dat = matrix_to_stack(mat, "value", "site", "species")
dat$site = as.character(dat$site)
dat$species = as.character(dat$species)

di_df = distinctiveness_stack(dat, "species", "site", "value", dist_mat)
head(di_df)

}

