% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relative.R
\name{make_absolute}
\alias{make_absolute}
\title{Abundance matrix from relative abundances matrix}
\usage{
make_absolute(rel_abund_matrix)
}
\arguments{
\item{rel_abund_matrix}{a relative abundance matrix (proportion of individuals
of a given species at a given site)}
}
\description{
Convert the input relative abundances matrix to an abundance matrix
}
\seealso{
\code{\link[funrar]{make_relative}} for the reverse operation
}

