% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relative.R
\name{make_relative}
\alias{make_relative}
\title{Relative abundance matrix from absolute abundance matrix}
\usage{
make_relative(abund_matrix)
}
\arguments{
\item{abund_matrix}{abundance matrix, with sites in rows and species in
columns.}
}
\value{
Similar shaped matrix as the input but with relative abundances
instead
}
\description{
From an abundance matrix (numbers of individuals of a given species at a
site) returns a relative abundance matrix (proportion of individuals of a
given species at a given site). This function works also
with sparse matrices.
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix
mat = as.matrix(aravo$spe)
head(mat)[, 1:5]  # Has absolute abundances
rel_mat = make_relative(mat)
head(rel_mat)  # Relative abundances

}
