% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_matrix.R
\name{matrix_to_tidy}
\alias{matrix_to_tidy}
\title{Matrix to tidy data.frame}
\usage{
matrix_to_tidy(my_mat, value_col = "value",
  row_to_col = names(dimnames(my_mat))[1],
  col_to_col = names(dimnames(my_mat))[2])
}
\arguments{
\item{my_mat}{matrix you want to transform in tidy data.frame}

\item{value_col}{(optional) character vector to use for value column (default
: 'value')}

\item{row_to_col}{(optional) character vector used for name of column in
data.frame corresponding to rows in matrix (default: corresponding
dimension name)}

\item{col_to_col}{(optional) character vector used for name of column in
data.frame corresponding to columns in matrix (default: corresponding
dimension name)}
}
\description{
From a matrix with values to a tidy data.frame
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe)
dat = matrix_to_tidy(mat, "value", "site", "species")
str(dat)

}

