% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctiveness.R
\name{pres_distinctiveness}
\alias{pres_distinctiveness}
\title{Distinctiveness on presence/absence matrix}
\usage{
pres_distinctiveness(pres_matrix, dist_mat)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix, with species in rows and sites
in columns (not containing relative abundances for the moments)}

\item{dist_matrix}{a functional distance matrix containing the species from
the provided presence-absence matrix}
}
\value{
a similar matrix to presence-absence with distinctiveness at sites?
}
\description{
Computes distinctiveness from a presence-absence matrix of species with a
provided functional distance matrix.
}
\details{
Experimental for the moment, should be merged with previous function
'distictinctiveness()'
}

