% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedness.R
\name{pres_restrictedness}
\alias{pres_restrictedness}
\title{Restrictedness on presence/absence matrix}
\usage{
pres_restrictedness(pres_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix, with sites in rows and species
in columns}
}
\value{
A tidy data.frame containing species' names and their restrictedness value
}
\description{
Computes restrictedness from a presence-absence matrix of species with a
provided functional distance matrix.
}

