% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarcity.R
\name{pres_scarcity}
\alias{pres_scarcity}
\title{Distinctiveness on presence/absence matrix}
\usage{
pres_scarcity(pres_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix, with species in rows and sites
in columns (not containing relative abundances for the moments)}
}
\value{
a similar matrix to presence-absence with scarcity values
}
\description{
Computes scarcity from a relative abundance matrix of species.
}
\details{
Experimental for the moment, should be merged with previous function
'scarcity()'
}

