% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{pres_uniqueness}
\alias{pres_uniqueness}
\title{Uniqueness for presence/Aasence matrix}
\usage{
pres_uniqueness(pres_matrix, dist_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix (or relative abundances),
with sites in rows and species in columns}

\item{dist_matrix}{a functional distance matrix}
}
\value{
a similar matrix to presence-absence with uniqueness values
}
\description{
Computes uniqueness from a presence-absence matrix of species with a
provided functional distance matrix. The
sites-species matrix should have \strong{sites} in \strong{rows} and
\strong{species} in \strong{columns}, similar to \code{\link[vegan]{vegan}}
package defaults.
}

