% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedness.R
\name{restrictedness}
\alias{restrictedness}
\title{Restrictedness on presence/absence matrix}
\usage{
restrictedness(pres_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix, with species in rows and sites
in columns (not containing relative abundances for the moments)}
}
\value{
A tidy data.frame containing species' names and their restrictedness value
}
\description{
Computes restrictedness from a presence-absence matrix of species with a
provided functional distance matrix.
}

