% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarcity.R
\name{scarcity}
\alias{scarcity}
\title{Scarcity}
\usage{
scarcity(com_table, species, com, abund)
}
\arguments{
\item{com_table}{a data frame of species in occurences in
communities.}

\item{species}{a character vector indicating the column of species in
\code{com_table}}

\item{com}{a character vector indicating the column name of communities ID in
\code{com_table}}

\item{abund}{a character vector indicating the column name of the relative
abundances of species in \code{com_table}}
}
\value{
The same table as \code{com_table} with an added \eqn{S_i} column
    for Scarcity values.
}
\description{
Compute scarcity values for several communities. Scarcity is computed per
community. Scarcity corresponds to the rareness of a given species in
terms of abundances, as such:
\deqn{
 S_i = \exp(-N\log{2}A_i),
}
with \eqn{S_i} the scarcity of species \eqn{i}, \eqn{N} the number of
species present in the given community and \eqn{A_i} the relative abundance
(in \%) of species \eqn{i}.
}

