% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseness.R
\name{sparseness}
\alias{sparseness}
\title{Sparseness}
\usage{
sparseness(com_table, species, com, abund)
}
\arguments{
\item{com_table}{a data frame a tidy version of species in occurences in
communities.}

\item{species}{a character vector indicating the column of species in
\code{com_table}}

\item{com}{a character vector indicating the column name of communities ID}

\item{abund}{a character vector indicating the column name of the relative
abundances of species}
}
\value{
The same table as \code{com_table} with an added \eqn{S_i} column
    for Sparseness values.
}
\description{
Compute sparness values for several communities. Sparseness is computed per
community. Sparseness corresponds to the rareness of a given species in
terms of abundances, as such:
\deqn{
 S_i = \exp(-N\log{2}A_i),
}
with \eqn{S_i} the sparseness of species \eqn{i}, \eqn{N} the number of
species present in the given community and \eqn{A_i} the relative abundance
(in %%) of species \eqn{i}.
}

