% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_matrix.R
\name{stack_to_matrix}
\alias{stack_to_matrix}
\title{Tidy data.frame to matrix}
\usage{
stack_to_matrix(my_df, col_to_row, col_to_col, col_value = NULL)
}
\arguments{
\item{my_df}{data.frame you want to transform in matrix}

\item{col_to_row}{character vector of the name of the data.frame column you
want to put into matrix rows}

\item{col_to_col}{character vector of the name of the data.frame column you
want to be as columns in matrix}

\item{col_value}{(optional) character vector indicating the name of a column
coding the values that will be put in the matrix}
}
\description{
Passes from a tidy data frame (only present species) with an index to a
matrix.
}
\examples{
example = data.frame("sites" = c(rep("1", 3), rep("2", 2)),
 "species" = c("A", "B", "C", "B", "D"),
  "abundance" = c(0.33, 0.33, 0.33, 0.4, 0.6))

mat = stack_to_matrix(example, "sites", "species", "abundance")
mat

}

