% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedness.R
\name{table_restrictedness}
\alias{table_restrictedness}
\title{Table Restrictedness}
\usage{
table_restrictedness(com_df, com, species)
}
\arguments{
\item{com_df}{a tidy data.frame of community with column with species}

\item{com}{a character vector indicating the name of community column}

\item{species}{a character vector indicating the name of species column}
}
\value{
A tidy data.frame containing species' names and their restrictedness
   value
}
\description{
Compute restrictedness for every species
}
\examples{
data("aravo", package = "ade4")
# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe); dat <- matrix_to_tidy(mat, "value", "site", "species")
dat$site = as.character(dat$site)
dat$species = as.character(dat$species)
ri_df = table_restrictedness(dat, "site", "species")
head(ri_df)

}

