% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarcity.R
\name{table_scarcity}
\alias{table_scarcity}
\title{Scarcity}
\usage{
table_scarcity(com_df, species, com, abund)
}
\arguments{
\item{com_df}{a data frame of species in occurences in
communities.}

\item{species}{a character vector indicating the column of species in
\code{com_df}}

\item{com}{a character vector indicating the column name of communities ID in
\code{com_df}}

\item{abund}{a character vector indicating the column name of the relative
abundances of species in \code{com_df}}
}
\value{
The same table as \code{com_df} with an added \eqn{S_i} column
    for Scarcity values.
}
\description{
Compute scarcity values for several communities. Scarcity is computed per
community. Scarcity corresponds to the rareness of a given species in
terms of abundances, as such:
\deqn{
 S_i = \exp(-N\log{2}A_i),
}{%
 S_i = exp^(-N*ln2*A_i),
}
with \eqn{S_i} the scarcity of species \eqn{i}, \eqn{N} the number of
species present in the given community and \eqn{A_i} the relative abundance
(in \%) of species \eqn{i}.
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe); dat <- matrix_to_tidy(mat, "value", "site", "species")
dat$site = as.character(dat$site)
dat$species = as.character(dat$species)

si_df = table_scarcity(dat, "species", "site", "value")
head(si_df)

}

