% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_matrix.R
\name{tidy_to_matrix}
\alias{tidy_to_matrix}
\title{Tidy data.frame to matrix}
\usage{
tidy_to_matrix(my_df, col_to_row, col_to_col, col_value = NULL)
}
\arguments{
\item{my_df}{data.frame you want to transform in matrix}

\item{col_to_row}{character vector of the name of the data.frame column you
want to put into matrix rows}

\item{col_to_col}{character vector of the name of the data.frame column you
want to be as columns in matrix}

\item{col_value}{(optional) character vector indicating the name of a column
coding the values that will be put in the matrix}
}
\description{
Passes from a tidy data frame (only present species) with an index to a
matrix.
}

