% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{uniqueness}
\alias{uniqueness}
\title{Uniqueness for presence/absence matrix}
\usage{
uniqueness(pres_matrix, dist_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix (or relative abundances),
with sites in rows and species in columns}

\item{dist_matrix}{a species functional distance matrix}
}
\value{
a data frame with uniqueness value per species
}
\description{
Computes uniqueness from a presence-absence matrix of species with a
provided functional distance matrix. The
sites-species matrix should have \strong{sites} in \strong{rows} and
\strong{species} in \strong{columns}, similar to \code{\link[vegan]{vegan}}
package defaults.
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix
mat = as.matrix(aravo$spe)
colnames(mat) = as.character(colnames(mat))

# Example of trait table
tra <- aravo$traits[, c("Height", "SLA", "N_mass")]
# Distance matrix
dist_mat <- compute_dist_matrix(tra)

ui = uniqueness(mat, dist_mat)
ui[1:5, 1:5]

}

