% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\alias{distance_matrix}
\title{Functional Distance Matrix}
\usage{
compute_dist_matrix(traits_table)
}
\arguments{
\item{traits_table}{A data frame of traits with species in row and traits in
columns, row names should be species names.}
}
\value{
A Gower's distance matrix (Gower's 1961), columns and rows get same
    names as \code{trait_table} row names. Be sure to give them consistent
    row names.
}
\description{
This function computes a Gower's Distance matrix of trait between each pair
of species present in given 'traits_table', each row represents a species
and each column a trait.
}
\examples{
set.seed(1)  # For reproducibility
trait = data.frame(sp = paste("sp", 1:5), trait_1 = runif(5),
   trait_2 = as.factor(c("A", "A", "A", "B", "B")))
rownames(trait) = trait$sp

dist_mat = compute_distance_matrix(trait[, -1])

}

