% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctiveness.R
\name{distinctiveness_com}
\alias{distinctiveness_com}
\title{Distinctiveness for a single community in stack data.frame}
\usage{
distinctiveness_com(com_df, sp_col, abund = NULL, dist_matrix)
}
\arguments{
\item{com_df}{a stacked (or tidy) data.frame from a single community with
each row representing a species}

\item{sp_col}{a character vector the name of the species column in
\code{com_df}}

\item{abund}{a character vector indicating the name of the column containing
relative abundances values}

\item{dist_matrix}{a functional distance matrix as given by
\code{compute_dist_matrix()}}
}
\value{
the same data.frame with the additional \strong{Di} column giving
   functional distinctiveness values for each species
}
\description{
Given a stacked data.frame and a distance matrix compute the functional
distinctiveness. See \code{\link[funrar]{distinctiveness}} function or the
functional rarity indices vignette included in the package (type
\code{vignette(package = "funrar")}).
}

