% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funrar.R
\name{funrar_stack}
\alias{funrar_stack}
\title{Functional Rarity indices for data.frame}
\usage{
funrar_stack(com_df, sp_col, com, abund = NULL, dist_matrix)
}
\arguments{
\item{com_df}{a data frame of the species in the regional pool.}

\item{sp_col}{a character vector indicating the name of the species column
in the \code{com_df} data frame}

\item{com}{a character vector, indicating the column name of communities
names.}

\item{abund}{optional character vector indicating the column name of the
relative abundances of species, if provided, weight distinctiveness by
relative abundances.}

\item{dist_matrix}{a functional distance matrix}
}
\value{
A list of 3 objects (or 4 if \code{rel_abund = TRUE}):
   \describe{
     \item{\strong{Ui}}{a vector containing uniqueness values per species,}
     \item{\strong{Di}}{a site-species matrix with functional distinctiveness
       values per species per site,}
     \item{\strong{Ri}}{a vector containing geographical restrictedness values
       per species,}
   }
   and if \code{rel_abund = TRUE},
   \describe{
     \item{\strong{Si}}{a site-species matrix with scarcity values per
       species per site.}
   }
}
\description{
From a stack community data.frame
}

