% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funrar.R
\name{funrar_stack}
\alias{funrar_stack}
\title{Functional Rarity indices for data.frame}
\usage{
funrar_stack(com_df, sp_col, com, abund = NULL, dist_matrix)
}
\arguments{
\item{com_df}{a stacked (or tidy) data.frame from a single community with
each row representing a species}

\item{sp_col}{a character vector the name of the species column in
\code{com_df}}

\item{com}{a character vector, indicating the column name of communities
names.}

\item{abund}{a character vector indicating the name of the column containing
relative abundances values}

\item{dist_matrix}{a functional distance matrix as given by
\code{compute_dist_matrix()}}
}
\value{
A list of 3 objects (or 4 if \code{rel_abund = TRUE}):
   \describe{
     \item{\strong{Ui}}{a vector containing uniqueness values per species,}
     \item{\strong{Di}}{a site-species matrix with functional distinctiveness
       values per species per site,}
     \item{\strong{Ri}}{a vector containing geographical restrictedness values
       per species,}
   }
   and if \code{rel_abund = TRUE},
   \describe{
     \item{\strong{Si}}{a site-species matrix with scarcity values per
       species per site.}
   }
}
\description{
From a stack community data.frame
}

