% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{pres_uniqueness}
\alias{pres_uniqueness}
\title{Uniqueness for Presence/Absence matrix}
\usage{
pres_uniqueness(pres_matrix, dist_matrix)
}
\arguments{
\item{pres_matrix}{a presence-absence matrix, with species in rows and sites
in columns (not containing relative abundances for the moment)}

\item{dist_matrix}{a functional distance matrix}
}
\value{
a similar matrix to presence-absence with uniqueness values
}
\description{
Computes uniqueness from a presence-absence matrix of species with a
provided functional distance matrix.
}
\details{
Experimental for the moment, should be merged with previous function
'uniqueness()'
}

