% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarcity.R
\name{scarcity_com}
\alias{scarcity_com}
\title{Scarcity for a single community in stack data.frame}
\usage{
scarcity_com(com_df, sp_col, abund)
}
\arguments{
\item{com_df}{a stacked (or tidy) data.frame from a single community with
each row representing a species}

\item{sp_col}{a character vector the name of the species column in
\code{com_df}}

\item{abund}{a character vector indicating the name of the column containing
relative abundances values}
}
\value{
the same data.frame with the additional \strong{Si} column giving
   scarcity values for each species
}
\description{
Given a stacked data.frame compute species scarcity.
See \code{\link[funrar]{scarcity}} function or the functional rarity indices
vignette included in the package (type \code{vignette(package = "funrar")})
for details about the indices.
}

