% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarcity.R
\name{scarcity_stack}
\alias{scarcity_stack}
\title{Scarcity}
\usage{
scarcity_stack(com_df, sp_col, com, abund)
}
\arguments{
\item{com_df}{a stacked (or tidy) data.frame from a single community with
each row representing a species}

\item{sp_col}{a character vector the name of the species column in
\code{com_df}}

\item{com}{a character vector indicating the column name of communities ID in
\code{com_df}}

\item{abund}{a character vector indicating the name of the column containing
relative abundances values}
}
\value{
The same table as \code{com_df} with an added \eqn{S_i} column
    for Scarcity values.
}
\description{
Compute scarcity values for several communities. Scarcity is computed per
community. Scarcity corresponds to the rareness of a given species in
terms of abundances, as such:
\deqn{
 S_i = \exp(-N\log{2}A_i),
}{%
 S_i = exp^(-N*ln2*A_i),
}
with \eqn{S_i} the scarcity of species \eqn{i}, \eqn{N} the number of
species present in the given community and \eqn{A_i} the relative abundance
(in \%) of species \eqn{i}.
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe); dat = matrix_to_stack(mat, "value", "site", "species")
dat$site = as.character(dat$site)
dat$species = as.character(dat$species)

si_df = scarcity_stack(dat, "species", "site", "value")
head(si_df)

}

