% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictedness.R
\name{table_restrictedness}
\alias{table_restrictedness}
\title{Table Restrictedness}
\usage{
table_restrictedness(com_table, com, species)
}
\arguments{
\item{com_table}{a tidy data.frame of community with column with species}

\item{com}{a character vector indicating the name of community column}

\item{species}{a character vector indicating the name of species column}
}
\value{
A tidy data.frame containing species' names and their restrictedness
   value
}
\description{
Compute restrictedness for every species
}

