% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{uniqueness_stack}
\alias{uniqueness_stack}
\title{Uniqueness}
\usage{
uniqueness_stack(com_df, sp_col, dist_matrix)
}
\arguments{
\item{com_df}{a data frame of the species in the regional pool.}

\item{sp_col}{a character vector indicating the name of the species column
in the \code{com_df} data frame}

\item{dist_matrix}{a functional distance matrix}
}
\value{
a data frame with uniqueness value per species
}
\description{
Computes Uniqueness values over a given regional pool.
The uniqueness value of a single species is equal to the
minimum functional distance with all other species in the regional pool. The
formula is as such:
\deqn{U_i = \min(d_{ij}),}{Ui = min(d_ij),}
with \eqn{d_{ij}}{d_ij} the functional distance between species
\eqn{i} and \eqn{j}.
}
\examples{
data("aravo", package = "ade4")

# Site-species matrix converted into data.frame
mat = as.matrix(aravo$spe); dat <- matrix_to_stack(mat, "value", "site", "species")
dat$site = as.character(dat$site)
dat$species = as.character(dat$species)

# Example of trait table
tra <- aravo$traits[, c("Height", "SLA", "N_mass")]
# Distance matrix
dist_mat <- compute_dist_matrix(tra)

ui_df = uniqueness_stack(dat, "species", dist_mat)
head(ui_df)

}

